properly-formatted counties string
