+++ frontend/src/components/InDemandTag.tsx
@@ -1,6 +1,7 @@
 import React, { ReactElement } from "react";
 import { useTranslation } from "react-i18next";
 import { InlineIcon } from "./InlineIcon";
+import { formatCountiesArrayToString } from "../utils/formatCountiesArrayToString";
 
 interface Props {
   className?: string;
@@ -10,22 +11,8 @@ interface Props {
 export const InDemandTag = (props: Props): ReactElement => {
   const { t } = useTranslation();
 
-// Helper function to format array into human-readable string
-  const formatCounties = (counties: string[] = []) => {
-    if(counties.length === 0) return '';
-    if(counties.length === 1) return `${counties[0]} County`;
-
-    const lastCountyIndex = counties.length - 1;
-    if(counties.length === 2) return `${counties[0]} and ${counties[lastCountyIndex]} Counties`;
-
-    const lastCounty = counties[lastCountyIndex];
-    const remainingCounties = counties.slice(0, lastCountyIndex);
-    return `${remainingCounties.join(', ')}, and ${lastCounty} Counties`;
-  };
-
-
   // Formatted counties string
-  const countiesString = formatCounties(props.counties);
+  const countiesString = formatCountiesArrayToString(props.counties);
 
   return (
       <span className={`fin fas bg-orange-tag tag pvxxs phd mrs ${props.className}`}>
